# swagger.yaml
swagger: '2.0'
info:
  title: Dendro Research Data Management Platform API
  description: Integrate your research data management workflows with Dendro **(Documentation IN CONSTRUCTION)**
  version: "0.2.0"
# the domain of the service
host: "127.0.0.1:3001"
...
basePath: "/"
...
schemes:
  - https
produces:
  - application/json
paths:
...
  /ontologies/public:
    get:
      summary: Provides a list of public ontologies in this Dendro
      description: |
        Lists the public ontologies loaded in this Dendro. A public ontology is one that can be used by users to describe datasets. A private ontology is used to represent private information (e.g the user password).
      tags:
        - Ontologies
      produces:
        ["text/json"]
      consumes:
        ["text/json"]
      responses:
        200:
          description: A list of ontologies
          schema:
            type: array
            items:
              $ref: '#/definitions/Ontology'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /ontologies/all:
    get:
      summary: Provides a list of ALL ontologies in this Dendro
      description: |
        Lists ALL the ontologies loaded in this Dendro, regardless of their visibility.
      tags:
        - Ontologies
      produces:
        ["text/json"]
      consumes:
        ["text/json"]
      responses:
        200:
          description: A list of ontologies
          schema:
            type: array
            items:
              $ref: '#/definitions/Ontology'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string
  Ontology:
    type: object
    properties:
      prefix:
        type: string
        description: The prefix of an ontology. For example, for Dublin Core Terms you have 'dcterms'
      uri:
        type: string
        description: The URI (in most cases also URL) that identifies the ontology uniquely.
      elements:
        type: array
        description: List of descriptors included in this ontology
        items:
          schema:
            type: array
            $ref: '#/definitions/Descriptor'
...
  Descriptor:
    type: object
    properties:
      uri:
        type: string
        format: uri
        description: The URI that identifies the descriptor
      prefix:
        type: string
        description: The prefix of the ontology to which the descriptor belongs
      ontology:
        type: string
        format: uri
        description: The URI of the ontology to which the descriptor belongs
      shortName:
        type: string
        format: uri
        description: The element name of the descriptor, i.e. what is to the right of the ':' in the prefixed form. E.g. for the 'dcterms:title' descriptor, it is 'title'
      prefixedForm:
        type: string
        format: uri
        description: The prefixed form of the identifier of this descriptor <prefix>:<shortName>
        #$ref: '#/definitions/TitleOnly'
      type:
        type: integer
        description: Determines the type of the descriptor, for serialization in the database. (for this value, just send back what Dendro sends you)
      control:
        type: string
        description: The type of control to present in the web interface, for this descriptor
      label:
        type: string
        description: Human-readable label for this descriptor. It is the value of its rdf:label property
      comment:
        type: string
        description: Human-readable comment for this descriptor. It is the value of its rdf:comment property
      recommendation_types :
        type: object
        description: When requesting descriptor recommendations, it provides an explanation of why the descriptor was recommended
      recommendationCallId:
        type: string
        format: uuid
        description: The ID of a recommendation call where the descriptor was recommended, if getting it from the descriptor recommendations endpoint. Send it back when recording an interaction so that Dendro knows from which recommendation session it was accepted or rejectec, for example. Important for feedback gathering on the effectiveness of the descriptor recommender.